+++ Algorithm/Descriptor/GLRLM.cs
@@ -30,7 +30,7 @@ namespace Baimp
 			int height = data.Height;
 			int width = data.Width;
 			int stride = data.Stride;
-			double[,] matrix = new double[256, width+1];
+			float[,] matrix = new float[256, width+1];
 
 			byte* src = (byte*) (data.Scan0);
 
@@ -69,7 +69,7 @@ namespace Baimp
 
 			//if (crop) { // make this an option?
 			if (maxRunLength < width+1) {
-				double[,] matrixTmp = new double[256, maxRunLength+1];
+				float[,] matrixTmp = new float[256, maxRunLength+1];
 				for (int y = 0; y < maxRunLength+1; y++ ) {
 					for (int x = 0; x < 255; x++) {
 						matrixTmp[x, y] = matrix[x, y];
